+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/util/FindbugsUtilsTest.groovy
@@ -1,35 +1,61 @@
 package ru.vyarus.gradle.plugin.quality.util
 
-import org.slf4j.Logger
-import org.slf4j.LoggerFactory
-import spock.lang.Specification
+import ru.vyarus.gradle.plugin.quality.AbstractTest
 
 import static ru.vyarus.gradle.plugin.quality.util.FindbugsUtils.mergeExcludes
-
 
 /**
  * @author Vyacheslav Rusakov
  * @since 20.03.2017
  */
-class FindbugsUtilsTest extends Specification {
-
-    Logger logger = LoggerFactory.getLogger(FindbugsUtilsTest.class)
+class FindbugsUtilsTest extends AbstractTest {
 
     def "Check exclude xml modification"() {
 
-        def sep = File.separator
-
         expect: "conversion"
-        mergeExcludes(['**/sample/*', '**/*.java'], exclude(), logger).text.replaceAll('\r', '') ==
-"""<?xml version="1.0" encoding="UTF-8"?><FindBugsFilter>
+        mergeExcludes(exclude(),
+                [file('src/main/java/sample/Sample.java'), file('src/main/java/other/Sample2.java')],
+                [file('/src/main/java')]
+        ).text.replaceAll('\r', '') ==
+                """<?xml version="1.0" encoding="UTF-8"?><FindBugsFilter>
   <Match>
     <Source name="~.*\\.groovy"/>
   </Match>
   <Match>
-    <Source name="~.*${sep}sample${sep}.*"/>
+    <Class name="sample.Sample"/>
   </Match>
   <Match>
-    <Source name="~.*${sep}.*\\.java\$"/>
+    <Class name="other.Sample2"/>
+  </Match>
+</FindBugsFilter>
+""" as String
+    }
+
+    def "Check no excludes"() {
+
+        expect: "no modification"
+        mergeExcludes(exclude(),
+                [],
+                [file('/src/main/java')]
+        ).text.replaceAll('\r', '') ==
+                """<?xml version="1.0" encoding="UTF-8"?><FindBugsFilter>
+  <Match>
+    <Source name="~.*\\.groovy"/>
+  </Match>
+</FindBugsFilter>
+""" as String
+    }
+
+    def "Check no matching roots"() {
+
+        expect: "no changes"
+        mergeExcludes(exclude(),
+                [file('src/main/java/sample/Sample.java'), file('src/main/java/other/Sample2.java')],
+                [file('/src/main/bad')]
+        ).text.replaceAll('\r', '') ==
+                """<?xml version="1.0" encoding="UTF-8"?><FindBugsFilter>
+  <Match>
+    <Source name="~.*\\.groovy"/>
   </Match>
 </FindBugsFilter>
 """ as String