+++ CHANGELOG.md
@@ -6,7 +6,7 @@
     check due to false positives
 * Add unified exclusions configuration: exclude extension property accepting glob values (#2)
     - Does not affect animalsniffer because it's a different type of check
-    - Findbugs did not support direct exclusions, so plugin converts blobs to regexps and adds them to exclusion xml (default or user defined)
+    - Findbugs did not support direct exclusions, so plugin resolves excluded classes and adds them to exclusion xml (default or user defined)
 
 ### 2.1.0 (2017-01-24)
 * Update checkstyle 7.1 -> 7.4